Elastic Image Recognition Application Developement/Web_Tier/requirements.txt
boto3==1.26.74
botocore==1.29.74
click==8.1.3
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
Pillow==9.4.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.13.0
